requirements.txt
pdfplumber
pdf2image
pytesseract
Pillow
google-generativeai
